+++ test/go_test.exs
@@ -1,8 +1,21 @@
 defmodule GoTest do
   use ExUnit.Case
-  doctest Go
+  # doctest Go
   
-  # test "the truth" do
-  #   assert 1 + 1 == 2
-  # end
+  alias Go.Game
+  
+  test "can create a new game" do
+    game = Go.new_game()
+    assert %Game{} = game
+  end
+  
+  test "can create a new game from a map" do
+    game = Go.new_game(%{size: 19})
+    assert %Game{size: 19} = game
+  end
+  
+  test "can create a new game from a keyword list" do
+    game = Go.new_game(size: 19)
+    assert %Game{size: 19} = game
+  end
 end